pyproject.toml
[project]
name = "youtube_channel"
version = "0.1.0"
description = "youtube-channel using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.203.1,<1.0.0",
    "ddgs>=9.6.1",
    "duckduckgo-search>=8.1.1",
    "elevenlabs>=2.19.0",
    "langchain>=1.0.1",
    "langchain-community>=0.4",
    "openai-whisper>=20250625",
    "google-genai>=0.3.3",
    "pillow>=10.4.0",
    "moviepy>=1.0.3",
    "playwright>=1.40.0",
    "anthropic>=0.18.0",
    "rembg>=2.0.50",
    "yfinance>=0.2.40",
    "matplotlib>=3.8.0",
]

[project.scripts]
youtube_channel = "src.main:run"
run_crew = "src.main:run"
train = "src.main:train"
replay = "src.main:replay"
test = "src.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.crewai]
type = "crew"
